Split resolver expr local tests
